requirements.txt
django
pyJWT
django-cors-headers
dj-database-url